backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pillow==10.3.0
numpy==1.26.4
sqlmodel==0.0.21
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0